+++ CHANGELOG.md
@@ -1,12 +1,18 @@
 # Later
 
+- [ ] Support Start forms (8.3)
+- [ ] Currently, Cross Origin request are allowed.
 - [ ] Able to register listeners for `assignment`, `delete`, `update`, and `timeout`
 - [ ] Integrate forms.js react styles (rather than simply adding styles to `index.html`)
 
 # 8.2.15-SNAPSHOT (October 2023)
 
 - [ ] Able to register complete task listener
+- [ ] Implement find task by variable
+- [ ] Able to complete task and load next form
 - [ ] Able to register create task listener
+- [x] Create sample bpmn process
+- [x] Able to start process instance from start form
 - [x] Able to display forms.js form
 - [x] Add sample react js app
 - [x] Implement `POST /process/start/{{processId}}`